jarvis-system/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.39.0
websockets
pyttsx3
SpeechRecognition
pywhatkit
wikipedia
requests
pyautogui
psutil
pyjokes
screen_brightness_control
plyer
opencv-python-headless==4.9.0.80
beautifulsoup4
pillow
pygetwindow
mss
pyfiglet
colorama
pyaudio
setuptools
pytz
textblob
scikit-learn
nltk
numpy
chromadb==0.4.22
sentence-transformers==2.3.1
pymupdf
langchain==0.3.28
langchain-core==0.3.83
langchain-community==0.3.31
langsmith==0.4.37
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
PyPDF2
python-docx
markdown
pandas
PyYAML
rake-nltk
yake
textstat
python-magic
pyperclip
watchdog
humanize
send2trash
pyzipper
mypy
python-pptx
aiohttp==3.11.12
matplotlib
pdf2image==1.17.0
pytesseract==0.3.10
shapely
pydantic-settings==2.11.0
sqlalchemy
tenacity
backoff
netifaces
GPUtil
distro
GitPython
anthropic
google-generativeai
openai
edge-tts
gTTS
rpa
applescript
unidecode
python-Levenshtein
rapidfuzz
sounddevice
pygame
jedi
radon
memory_profiler
line_profiler
snakeviz
python-multipart
aiofiles
cryptography
passlib
python-jose
jedi
pydocstyle
yapf
autopep8
isort
pylint
pyflakes
pycodestyle
snowballstemmer
termcolor
art
prettytable
rich
rich-cli
yaspin
alive-progress
inquirer
editor
readchar
blessed
graphemeu
about-time
runs
xmod
mando

